Refactoring new-components into components
